Module rustfmt::config

Structs

Config

ConfigHelpItem

ConfigSetter

PartialConfig

Enums

BraceStyle

ControlBraceStyle

Density

IndentStyle

LicensePolicy

MultilineStyle

NewlineStyle

ReportTactic

ReturnIndent

Style

TypeDensity

WriteMode

Traits

ConfigType
Trait for types that can be used in Config.

Functions

get_toml_path
Check for the presence of known config file names (rustfmt.toml,.rustfmt.toml) indir`