Struct rustfmt::config::Config

pub struct Config { /* fields omitted */ }

Methods

impl Config

pub fn verbose(&self) -> bool

pub fn disable_all_formatting(&self) -> bool

pub fn skip_children(&self) -> bool

pub fn file_lines(&self) -> FileLines

pub fn max_width(&self) -> usize

pub fn error_on_line_overflow(&self) -> bool

pub fn tab_spaces(&self) -> usize

pub fn fn_call_width(&self) -> usize

pub fn struct_lit_width(&self) -> usize

pub fn struct_variant_width(&self) -> usize

pub fn force_explicit_abi(&self) -> bool

pub fn newline_style(&self) -> NewlineStyle

pub fn fn_brace_style(&self) -> BraceStyle

pub fn item_brace_style(&self) -> BraceStyle

pub fn control_style(&self) -> Style

pub fn control_brace_style(&self) -> ControlBraceStyle

pub fn impl_empty_single_line(&self) -> bool

pub fn trailing_comma(&self) -> SeparatorTactic

pub fn fn_empty_single_line(&self) -> bool

pub fn fn_single_line(&self) -> bool

pub fn fn_return_indent(&self) -> ReturnIndent

pub fn fn_args_paren_newline(&self) -> bool

pub fn fn_args_density(&self) -> Density

pub fn fn_args_layout(&self) -> IndentStyle

pub fn array_layout(&self) -> IndentStyle

pub fn array_width(&self) -> usize

pub fn type_punctuation_density(&self) -> TypeDensity

pub fn where_style(&self) -> Style

pub fn where_density(&self) -> Density

pub fn where_layout(&self) -> ListTactic

pub fn where_pred_indent(&self) -> IndentStyle

pub fn generics_indent(&self) -> IndentStyle

pub fn struct_lit_style(&self) -> IndentStyle

pub fn struct_lit_multiline_style(&self) -> MultilineStyle

pub fn fn_call_style(&self) -> IndentStyle

pub fn report_todo(&self) -> ReportTactic

pub fn report_fixme(&self) -> ReportTactic

pub fn chain_indent(&self) -> IndentStyle

pub fn chain_one_line_max(&self) -> usize

pub fn chain_split_single_child(&self) -> bool

pub fn reorder_imports(&self) -> bool

pub fn reorder_imports_in_group(&self) -> bool

pub fn reorder_imported_names(&self) -> bool

pub fn single_line_if_else_max_width(&self) -> usize

pub fn format_strings(&self) -> bool

pub fn force_format_strings(&self) -> bool

pub fn take_source_hints(&self) -> bool

pub fn hard_tabs(&self) -> bool

pub fn wrap_comments(&self) -> bool

pub fn comment_width(&self) -> usize

pub fn normalize_comments(&self) -> bool

pub fn wrap_match_arms(&self) -> bool

pub fn match_block_trailing_comma(&self) -> bool

pub fn indent_match_arms(&self) -> bool

pub fn closure_block_indent_threshold(&self) -> isize

pub fn space_before_type_annotation(&self) -> bool

pub fn space_after_type_annotation_colon(&self) -> bool

pub fn space_before_struct_lit_field_colon(&self) -> bool

pub fn space_after_struct_lit_field_colon(&self) -> bool

pub fn space_before_bound(&self) -> bool

pub fn space_after_bound_colon(&self) -> bool

pub fn spaces_around_ranges(&self) -> bool

pub fn spaces_within_angle_brackets(&self) -> bool

pub fn spaces_within_square_brackets(&self) -> bool

pub fn spaces_within_parens(&self) -> bool

pub fn use_try_shorthand(&self) -> bool

pub fn write_mode(&self) -> WriteMode

pub fn condense_wildcard_suffixes(&self) -> bool

pub fn combine_control_expr(&self) -> bool

pub fn set<'a>(&'a mut self) -> ConfigSetter<'a>

pub fn from_toml(toml: &str) -> Result<Config, String>

pub fn used_options(&self) -> PartialConfig

pub fn all_options(&self) -> PartialConfig

pub fn override_value(&mut self, key: &str, val: &str)

pub fn from_toml_path(file_path: &Path) -> Result<Config, Error>

Construct a Config from the toml file specified at file_path.

This method only looks at the provided path, for a method that searches parents for a rustfmt.toml see from_resolved_toml_path.

Return a Config if the config could be read and parsed from the file, Error otherwise.

pub fn from_resolved_toml_path(
dir: &Path
) -> Result<(Config, Option<PathBuf>), Error>

Resolve the config for input in dir.

Searches for rustfmt.toml beginning with dir, and recursively checking parents of dir if no config file is found. If no config file exists in dir or in any parent, a default Config will be returned (and the returned path will be empty).

Returns the Config to use, and the path of the project file if there was one.

pub fn print_docs()