Struct rustfmt::file_lines::FileLines

pub struct FileLines(_);

A set of lines in files.

It is represented as a multimap keyed on file names, with values a collection of non-overlapping ranges sorted by their start point. An inner None is interpreted to mean all lines in all files.

Methods

impl FileLines

pub fn all() -> FileLines

Creates a FileLines that contains all lines in all files.

Important traits for Files<'a>
impl<'a> Iterator for Files<'a> type Item = &'a String;

pub fn files(&self) -> Files

Returns an iterator over the files contained in self.

pub fn contains(&self, range: &LineRange) -> bool

Returns true if range is fully contained in self.

pub fn intersects(&self, range: &LineRange) -> bool

Returns true if any lines in range are in self.

pub fn contains_line(&self, file_name: &str, line: usize) -> bool

Returns true if line from file_name is in self.

pub fn intersects_range(&self, file_name: &str, lo: usize, hi: usize) -> bool

Returns true if any of the lines between lo and hi from file_name are in self.