Struct rustfmt::Shape

pub struct Shape {
    pub width: usize,
    pub indent: Indent,
    pub offset: usize,
}

Fields
width: usize indent: Indent offset: usize
Methods

impl Shape

pub fn legacy(width: usize, indent: Indent) -> Shape

indent is the indentation of the first line. The next lines should begin with at least indent spaces (except backwards indentation). The first line should not begin with indentation. width is the maximum number of characters on the last line (excluding indent). The width of other lines is not limited by width. Note that in reality, we sometimes use width for lines other than the last (i.e., we are conservative).

pub fn indented(indent: Indent, config: &Config) -> Shape

pub fn with_max_width(&self, config: &Config) -> Shape

pub fn offset(width: usize, indent: Indent, offset: usize) -> Shape

pub fn visual_indent(&self, extra_width: usize) -> Shape

pub fn block_indent(&self, extra_width: usize) -> Shape

pub fn block_left(&self, width: usize) -> Option<Shape>

pub fn add_offset(&self, extra_width: usize) -> Shape

pub fn block(&self) -> Shape

pub fn sub_width(&self, width: usize) -> Option<Shape>

pub fn shrink_left(&self, width: usize) -> Option<Shape>

pub fn offset_left(&self, width: usize) -> Option<Shape>

pub fn used_width(&self) -> usize