Crate rustfmt

Modules

codemap
This module contains utilities that work with the CodeMap from libsyntax / syntex_syntax. This includes extension traits and methods for looking up spans and line ranges for AST nodes.

config

file_lines
This module contains types and functions to support formatting specific line ranges.

filemap

modules

rustfmt_diff

visitor

Macros

impl_enum_decodable

try_opt

Structs

FormatReport

FormattingError

Indent

Summary

Enums

ErrorKind

Input

Traits

Spanned

Functions

format_input

run