Struct rustfmt::file_lines::FileLines

pub struct FileLines(_);

A set of lines in files.

It is represented as a multimap keyed on file names, with values a collection of non-overlapping ranges sorted by their start point. An inner None is interpreted to mean all lines in all files.

Methods

impl FileLines

fn all() -> FileLines

Creates a FileLines that contains all lines in all files.

fn files(&self) -> Files

Returns an iterator over the files contained in self.

fn contains(&self, range: &LineRange) -> bool

Returns true if range is fully contained in self.

fn intersects(&self, range: &LineRange) -> bool

Returns true if any lines in range are in self.